# Use size hint for maps in findInteractionPoints