--- a/geom/dcel_interaction_points.go
+++ b/geom/dcel_interaction_points.go
@@ -10,14 +10,18 @@ package geom
 // - Input geometries don't have any repeated coordinates (e.g. in areal rings
 //   or linear elements).
 func findInteractionPoints(gs []Geometry) map[XY]struct{} {
-	interactions := make(map[XY]struct{})
+	var sizeHint int
+	for _, g := range gs {
+		sizeHint += g.controlPoints()
+	}
+	interactions := make(map[XY]struct{}, sizeHint)
 
 	// adjacents tracks the next and previous points relative to a middle point
 	// for linear elements (i.e. the points adjacents to a middle point). It is
 	// used to differentiate the cases where linear elements overlap (in which
 	// case there ISN'T an interaction point) and cases where they are crossing
 	// over each other (in which case there IS an interaction point).
-	adjacents := make(map[XY]xyPair)
+	adjacents := make(map[XY]xyPair, sizeHint)
 
 	for _, g := range gs {
 		addGeometryInteractions(g, adjacents, interactions)