## Description

Initiates the maps in findInteractionPoints with a size hint to reduce allocation and map resize overhead. Only gives a small perf improvement.

## Check List

Have you:

- Added unit tests? N/A, relies on existing.

- Add cmprefimpl tests? (if appropriate?) N/A, relies on existing.

## Related Issue

N/A

## Benchmark Results

```
COMPARISON
name                    old time/op    new time/op    delta
Intersection/n=10-4       54.7µs ± 5%    54.2µs ± 7%    ~     (p=0.170 n=13+15)
Intersection/n=100-4       309µs ±14%     284µs ± 6%  -7.99%  (p=0.001 n=15+14)
Intersection/n=1000-4     2.90ms ± 9%    2.72ms ±13%  -6.03%  (p=0.004 n=14+15)
Intersection/n=10000-4    29.5ms ±14%    28.8ms ±13%    ~     (p=0.512 n=15+15)

name                    old alloc/op   new alloc/op   delta
Intersection/n=10-4       29.2kB ± 0%    30.5kB ± 0%  +4.50%  (p=0.000 n=15+14)
Intersection/n=100-4       155kB ± 0%     145kB ± 0%  -6.28%  (p=0.000 n=15+13)
Intersection/n=1000-4     1.88MB ± 0%    1.71MB ± 0%  -9.00%  (p=0.000 n=15+15)
Intersection/n=10000-4    19.3MB ± 0%    17.9MB ± 0%  -7.13%  (p=0.000 n=15+15)

name                    old allocs/op  new allocs/op  delta
Intersection/n=10-4          284 ± 0%       281 ± 0%  -1.06%  (p=0.000 n=15+15)
Intersection/n=100-4         457 ± 0%       441 ± 0%  -3.50%  (p=0.000 n=15+15)
Intersection/n=1000-4      2.07k ± 0%     1.97k ± 0%  -4.55%  (p=0.000 n=15+15)
Intersection/n=10000-4     19.1k ± 0%     18.5k ± 0%  -3.21%  (p=0.000 n=15+15)
```